styling update authentication pages
