+++ src/App.tsx
@@ -4,10 +4,10 @@ import { ThemeProvider } from 'styled-components';
 import { useCookies } from 'react-cookie';
 import { darkTheme } from './theme';
 import { GlobalStyle } from './global-style';
-import SignUp from './pages/SignUp';
-import SignIn from './pages/SignIn';
-import ForgotPassword from './pages/ForgotPassword';
-import ResetPassword from './pages/ResetPassword';
+import SignUp from './pages/Authentication/SignUp';
+import SignIn from './pages/Authentication/SignIn';
+import ForgotPassword from './pages/Authentication/ForgotPassword';
+import ResetPassword from './pages/Authentication/ResetPassword';
 import MemberLayout from './pages/Member/MemberLayout';
 import Home from './pages/Home/Home';
 import EditorPlans from './pages/Editor/EditorPlans';
