+++ src/pages/Authentication/components/ThirdPartyLoginButton.tsx
@@ -1,0 +1,30 @@
+import React from 'react';
+import styled from 'styled-components';
+import { FaGithub } from 'react-icons/fa';
+import { Button } from 'components';
+import { GITHUB_AUTH_URL } from 'utilities/constants';
+
+export const GithubIcon = styled(FaGithub)`
+  font-size: 1.2rem;
+  color: ${({ theme }) => theme.color.primary};
+`;
+
+export const ThirdPartyLoginButton = styled(Button)`
+  border: 1px solid ${({ theme }) => theme.color.primary};
+  border-radius: 4px;
+  height: 2.5em;
+  color: ${({ theme }) => theme.color.primary};
+  width: 90%;
+`;
+
+export const GithubLoginButton = () => {
+  const handleGithubClick = () => {
+    window.open(GITHUB_AUTH_URL);
+  };
+
+  return (
+    <ThirdPartyLoginButton onClick={handleGithubClick}>
+      <GithubIcon /> Continue with Github
+    </ThirdPartyLoginButton>
+  );
+};
