+++ tsconfig.json
@@ -2,6 +2,7 @@
   "compilerOptions": {
     "target": "es5",
     "lib": ["dom", "dom.iterable", "esnext"],
+    "baseUrl": "src",
     "allowJs": true,
     "skipLibCheck": true,
     "esModuleInterop": true,
@@ -17,5 +18,5 @@
     "noImplicitAny": false,
     "jsx": "react-jsx"
   },
-  "include": ["src", "src/styles/AuthButton.tsx"]
+  "include": ["src/**/*"]
 }
